VSR: Change cluster id from u32 to u128
Rationale (via https://github.com/tigerbeetle/tigerbeetle/pull/1295#discussion_r1389061252):

> A 128-bit cluster also improves our "resolution" for interoperability with other systems that want to use UUIDs for cluster ID to allow for random generation of cluster ID without the cost of having to introduce a centralized ID coordinator. Plus it gives us a further security guarantee that clusters can't talk to other clusters unless they can literally “guess” the identifier, which is now cryptographically "tough" (granted, we might not be able to imagine now why this could be important, but it's a stronger property nevertheless—especially from an online verification point of view).

+++ src/benchmark.zig
@@ -105,7 +105,7 @@ pub fn main() !void {
     );
 
     const client_id = std.crypto.random.int(u128);
-    const cluster_id: u32 = 0;
+    const cluster_id: u128 = 0;
 
     var io = try IO.init(32, 0);
